+++ packages/account-connector/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sophon-labs/account-connector
 
+## 3.0.0-next.16
+
+### Patch Changes
+
+- Previewing sophon account address off chain
+- Updated dependencies
+  - @sophon-labs/account-provider@3.0.0-next.16
+  - @sophon-labs/account-core@3.0.0-next.16
+
 ## 3.0.0-next.15
 
 ### Patch Changes